Katcam_pro/requirements.txt
tzlocal>=5,<6
tzdata>=2024.1
Pillow>=10
opencv-python>=4.9
